+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Pkg v1.11 Release Notes
 
 - Pkg now obeys `[compat]` bounds for `julia` and raises an error if the version of the running Julia binary is incompatible with the bounds in `Project.toml`.
   Pkg has always obeyed this compat when working with Registry packages. This change affects mostly local packages. ([#3526])
+- `pkg> add` and `Pkg.add` will now add compat entries for new direct dependencies if the active environment is a
+  package (has a `name` and `uuid` entry) ([#3732])
 
 Pkg v1.10 Release Notes
 =======================